+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - cmsdev: Changed default log directory to `/opt/cray/tests/install/logs/cmsdev/` to be consistent with other CSM tests.
 - cmsdev: Simplified Kubernetes artifact collection functions; collect additional information
 - cmsdev: Collect Kubernetes artifacts on failure by default
+- cmsdev: Compress test artifacts, if collected. If none collected, delete empty artifact directory, if cmsdev created it
 
 ### Removed